Services/ApiGateway/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
flask-marshmallow
py_eureka_client